requirements.txt
exifread
PyPDF2
python-docx
pandas
folium
